requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
playwright>=1.40.0
watchdog>=3.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
schedule>=1.2.0
